app/requirements.txt
anyio==3.3.0
appdirs==1.4.4
appnope==0.1.2
arrow==0.17.0
asgiref==3.4.1
astor==0.8.1
autopep8==1.5.7
backcall==0.2.0
beautifulsoup4==4.9.3
binaryornot==0.4.4
bs4==0.0.1
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
cookiecutter==1.7.2
cycler==0.10.0
decorator==5.0.9
devtools==0.6.1
distlib==0.3.2
elasticsearch==7.14.0
fastapi==0.67.0
filelock==3.0.12
h11==0.12.0
httpcore==0.13.6
httptools==0.2.0
httpx==0.18.2
idna==2.10
ipython==7.25.0
ipython-genutils==0.2.0
jedi==0.18.0
Jinja2==2.11.2
jinja2-time==0.2.0
kiwisolver==1.3.1
lxml==4.6.3
MarkupSafe==1.1.1
matplotlib==3.4.2
matplotlib-inline==0.1.2
multitasking==0.0.9
mypy==0.910
mypy-extensions==0.4.3
numpy==1.21.1
pandas==1.3.1
pandasgui==0.2.12
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.3.1
pipenv==2021.5.29
plotly==5.1.0
poyo==0.5.0
prompt-toolkit==3.0.19
ptyprocess==0.7.0
pyactiveresource==2.2.2
pyarrow==5.0.0
pycodestyle==2.7.0
pydantic==1.8.2
Pygments==2.9.0
PyJWT==2.1.0
pynput==1.7.3
pyobjc-core==7.3
pyobjc-framework-Cocoa==7.3
pyobjc-framework-Quartz==7.3
pyparsing==2.4.7
PyQt5==5.15.4
PyQt5-Qt5==5.15.2
PyQt5-sip==12.9.0
PyQtWebEngine==5.15.4
PyQtWebEngine-Qt5==5.15.2
python-dateutil==2.8.1
python-dotenv==0.19.0
python-slugify==4.0.1
pytz==2021.1
PyYAML==5.4.1
qtstylish==0.1.3
requests==2.25.1
rfc3986==1.5.0
ShopifyAPI==8.4.2
six==1.15.0
sniffio==1.2.0
soupsieve==2.2.1
starlette==0.14.2
tenacity==8.0.1
text-unidecode==1.3
toml==0.10.2
traitlets==5.0.5
typing-extensions==3.10.0.0
urllib3==1.26.3
uvicorn==0.14.0
uvloop==0.15.3
virtualenv==20.4.7
virtualenv-clone==0.5.4
watchgod==0.7
wcwidth==0.2.5
websockets==9.1
wordcloud==1.8.1
